src/test/java/org/alan/algorithm/practice/baekjoon/clazz/five/BJ9466Test.java
package org.alan.algorithm.practice.baekjoon.clazz.five;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

public class BJ9466Test {
    static Stream<Arguments> parameters() {
        return Stream.of(
                Arguments.of("""
                        7
                        3 1 3 7 3 4 6""", "3"),
                Arguments.of("""
                        8
                        1 2 3 4 5 6 7 8""", "0")
        );
    }
    
    @ParameterizedTest
    @MethodSource("parameters")
    public void test1(String input, String expected) {
        Assertions.assertEquals(expected, new BJ9466().solve2(input));
    }
}
